// src/app/models/achat/tiers.model.ts
import { TierCategory } from './../sales/tier-category.model';
import { Resource } from '../shared/ressource.model';
import { Contact } from '../shared/contact.model';
import { Currency } from '../administration/currency.model';
import { FileInfo } from '../shared/objectToSend';
import NumberFormatOptions = Intl.NumberFormatOptions;


export class Tiers extends Resource {
  IdPaymentCondition: number;
  IdTypeTiers: number;
  IdTierCategory: number;
  IdTierCategoryNavigation: TierCategory;
  IdCity: number;
  IdPays: number;
  Name: string;
  Adress: string;
  Region: string;
  CodeTiers: string;
  Status: number;
  Rib: string;
  Discount: number;
  IdCurrency: number;
  IdTaxeGroupTiersNavigation;
  Contact: Contact[];
  IdAccountingAccountTiers: number;
  IdCurrencyNavigation: Currency;
  AuthorizedAmountInvoice: number;
  PaymentDelay: number;
  IsCash?: boolean;
  ProvisionalAuthorizedAmountDelivery?: number;
  Email: string;
  Phone: string;
  IdPhone: number;
  IdPhoneNavigation: any;
  Fax: string;
  Linkedin: string;
  Facebook: string;
  Twitter: string;
  Description: string;
  Address: any[] = [];
  IdTaxeGroupTiers: number;
  DateToRemember: any[] = [];
  UrlPicture: string;
  PictureFileInfo: FileInfo;
  TiersPrices: any[] = [];
  Adresse: any[] = [];
  FilesInfos: Array<FileInfo>;
  IdSettlementMode: number;
  IdDeliveryType: number ;
  ActivitySector: string;
  FormatOption : NumberFormatOptions;
  TiersClassification?: number;
  image: any;
  WasLead: boolean;
  IdSalesPrice: number ;
  Vehicle: any[] = [];
  UpdatedDate: Date;
  CreationDate: Date;
  IsSynchronizedBToB: boolean;
}