// src/app/models/payment/settlement.model.ts
import { Resource } from '../shared/ressource.model';
import { DocumentStatus } from '../sales/document-status.model';
import { SettlementCommitment } from './settlement-commitment.model';
import { Tiers } from '../achat/tiers.model';
import { Currency } from '../administration/currency.model';
import { PaymentMethod } from '../payment-method/payment-method.model';
import { FileInfo } from '../shared/objectToSend';
import { SettlementType } from '../enumerators/settlement-type.enum';
import { BankAccount } from '../shared/bank-account.model';
import { PaymentStatus } from './payment-status.model';
import { ReflectiveSettlement } from './reflective-settlement.model';
import { PaymentSlip } from '../treasury/payment-slip.model';
import { Reconciliation } from '../treasury/reconciliation.model';
import { CashRegister } from '../treasury/cash-register.model';

export class Settlement extends Resource {
  Code: string;
  SettlementReference: string;
  IdTiers: number;
  Note: string;
  Holder: string;
  Type: SettlementType;
  Label: string;
  SettlementDate: Date;
  CommitmentDate: Date;
  IdPaymentMethod: number;
  Amount?: number;
  AmountWithCurrency: number;
  IdUsedCurrency?: number;
  ExchangeRate?: number;
  IdStatus?: number;
  IssuingBank: string;
  IdBankAccount?: number;
  Direction: number;
  Handed: string;
  IdPaymentStatus?: number;
  HasBeenReplaced: boolean;
  IdPaymentSlip: number;
  Checked: boolean;
  DeletedToken: string;
  IdPaymentSlipNavigation: PaymentSlip;
  IdBankAccountNavigation: BankAccount;
  IdPaymentMethodNavigation: PaymentMethod;
  IdPaymentStatusNavigation: PaymentStatus;
  IdStatusNavigation: DocumentStatus;
  IdTiersNavigation: Tiers;
  IdUsedCurrencyNavigation: Currency;
  IdReconciliation?: number;
  ReflectiveSettlementIdSettlementNavigation: Array<ReflectiveSettlement>;
  ReflectiveSettlementIdSettlementReplacedNavigation: Array<ReflectiveSettlement>;
  SettlementCommitment: Array<SettlementCommitment>;
  AttachmentUrl: string;
  WithholdingTax: number;
  WithholdingTaxAttachmentUrl: string;
  IdSessionCash? : number;
  Files: Array<any>;
  FilesInfos: Array<FileInfo>;
  UploadedFiles: Array<string>;
  ObservationsFilesInfo: Array<FileInfo>;
  WithholdingTaxObservationsFilesInfo: Array<FileInfo>;
  IdReconciliationNavigation: Reconciliation;
  IdSessionCashNavigation: CashRegister;
  IsDepositSettlement? : boolean;
}